millipixels google oauth 2/requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
requests==2.31.0
python-dotenv==1.0.0
